consolidate modules
